+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Kotlin language version handling in Android ([#3436](https://github.com/getsentry/sentry-dart/pull/3436))
 
+### Enhancements
+
+- Replace log batcher with telemetry processor ([#3448](https://github.com/getsentry/sentry-dart/pull/3448))
+
 ### Dependencies
 
 - Bump Native SDK from v0.10.0 to v0.12.3 ([#3438](https://github.com/getsentry/sentry-dart/pull/3438))